# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /context/brief:
    get:
      operationId: getContextBrief
      tags: [context]
      summary: Get personalized context brief
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [backend, frontend, pm]
          required: false
      responses:
        "200":
          description: Context brief for the given role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextBrief"
  /context/team:
    get:
      operationId: getContextTeam
      tags: [context]
      summary: Get recommended team members
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [backend, frontend, pm]
          required: false
      responses:
        "200":
          description: List of relevant team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
  /context/documents:
    get:
      operationId: getContextDocuments
      tags: [context]
      summary: Get AI-recommended documents
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [backend, frontend, pm]
          required: false
      responses:
        "200":
          description: List of recommended documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContextDocument"
  /openai/conversations:
    post:
      operationId: createConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /openai/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [openai]
      summary: Get messages for a conversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: sendMessage
      tags: [openai]
      summary: Send a message (SSE streaming response)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: Streamed AI response (text/event-stream)
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          type: string
          enum: [backend, frontend, pm]
      required:
        - email
        - password
        - name
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
    ContextBrief:
      type: object
      properties:
        keyInsight:
          type: string
        nextAction:
          type: string
        whyMatters:
          type: string
        confidence:
          type: integer
        suggestedPerson:
          type: object
          properties:
            name:
              type: string
            role:
              type: string
          required:
            - name
            - role
        reasoning:
          type: array
          items:
            type: string
      required:
        - keyInsight
        - nextAction
        - whyMatters
        - confidence
        - suggestedPerson
        - reasoning
    TeamMember:
      type: object
      properties:
        initials:
          type: string
        name:
          type: string
        role:
          type: string
        color:
          type: string
        tags:
          type: array
          items:
            type: string
        why:
          type: string
      required:
        - initials
        - name
        - role
        - color
        - tags
        - why
    ContextDocument:
      type: object
      properties:
        name:
          type: string
        why:
          type: string
      required:
        - name
        - why
    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    ConversationResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - createdAt
    MessageResponse:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        role:
          type: string
          enum: [backend, frontend, pm]
      required:
        - content
...