# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 8;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  name: zod.string(),
  role: zod.enum(["backend", "frontend", "pm"]).optional(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  role: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  role: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get personalized context brief
 */
export const GetContextBriefQueryParams = zod.object({
  role: zod.enum(["backend", "frontend", "pm"]).optional(),
});

export const GetContextBriefResponse = zod.object({
  keyInsight: zod.string(),
  nextAction: zod.string(),
  whyMatters: zod.string(),
  confidence: zod.number(),
  suggestedPerson: zod.object({
    name: zod.string(),
    role: zod.string(),
  }),
  reasoning: zod.array(zod.string()),
});

/**
 * @summary Get recommended team members
 */
export const GetContextTeamQueryParams = zod.object({
  role: zod.enum(["backend", "frontend", "pm"]).optional(),
});

export const GetContextTeamResponseItem = zod.object({
  initials: zod.string(),
  name: zod.string(),
  role: zod.string(),
  color: zod.string(),
  tags: zod.array(zod.string()),
  why: zod.string(),
});
export const GetContextTeamResponse = zod.array(GetContextTeamResponseItem);

/**
 * @summary Get AI-recommended documents
 */
export const GetContextDocumentsQueryParams = zod.object({
  role: zod.enum(["backend", "frontend", "pm"]).optional(),
});

export const GetContextDocumentsResponseItem = zod.object({
  name: zod.string(),
  why: zod.string(),
});
export const GetContextDocumentsResponse = zod.array(
  GetContextDocumentsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get messages for a conversation
 */
export const GetConversationMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const GetConversationMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetConversationMessagesResponse = zod.array(
  GetConversationMessagesResponseItem,
);

/**
 * @summary Send a message (SSE streaming response)
 */
export const SendMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
  role: zod.enum(["backend", "frontend", "pm"]).optional(),
});

/**
 * @summary Get overview summary
 */
export const GetInsightsSummaryResponse = zod.object({
  success: zod.boolean(),
  data: zod.object({}).passthrough(),
});

/**
 * @summary Get service risks
 */
export const GetInsightsRisksResponse = zod.object({
  success: zod.boolean(),
  data: zod.object({}).passthrough(),
});

/**
 * @summary Get topology graph
 */
export const GetInsightsGraphResponse = zod.object({
  success: zod.boolean(),
  data: zod.object({}).passthrough(),
});

/**
 * @summary Get latest root cause analysis
 */
export const GetInsightsRootCauseResponse = zod.object({
  success: zod.boolean(),
  data: zod.object({}).passthrough().optional(),
});

/**
 * @summary Ingest webhook alert
 */
export const IngestWebhookBody = zod.object({
  service: zod.string(),
  event: zod.string(),
  severity: zod.string().optional(),
  message: zod.string().optional(),
});

export const IngestWebhookResponse = zod.object({
  ok: zod.boolean(),
});